+++ services/__init__.py
@@ -1,1 +1,2 @@
 from .file_service import *
+from .tyre_impression_service import TyreImpressionService